fix: shift model events subscription to cortex
