+++ extensions/inference-cortex-extension/src/@types/global.d.ts
@@ -1,5 +1,6 @@
 declare const NODE: string
 declare const CORTEX_API_URL: string
+declare const CORTEX_SOCKET_URL: string
 declare const DEFAULT_SETTINGS: Array<any>
 declare const MODELS: Array<any>
